+++ src/main/java/at/hannibal2/skyhanni/events/LorenzToolTipEvent.kt
@@ -1,0 +1,6 @@
+package at.hannibal2.skyhanni.events
+
+import net.minecraft.inventory.Slot
+import net.minecraft.item.ItemStack
+
+class LorenzToolTipEvent(val slot: Slot, val itemStack: ItemStack, var toolTip: MutableList<String>): LorenzEvent()